Make full_simplify not be pub
